FractionTokenのテストコードをコントラクトに対応
