+++ pkgs/contract/gas-report.txt
@@ -7,25 +7,7 @@
 ··············|··········|··············|·············|·············|···············|··············
 |  Deployments           ·                                          ·  % of limit   ·             │
 ·························|··············|·············|·············|···············|··············
-|  BigBang               ·           -  ·          -  ·    1248128  ·        4.2 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  FractionToken         ·           -  ·          -  ·    2859199  ·        9.5 %  ·          -  │
+|  FractionToken         ·           -  ·          -  ·    3076626  ·       10.3 %  ·          -  │
 ·························|··············|·············|·············|···············|··············
 |  Hats                  ·           -  ·          -  ·    7032431  ·       23.4 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  HatsModule            ·           -  ·          -  ·     754132  ·        2.5 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  HatsModuleFactory     ·           -  ·          -  ·    1101122  ·        3.7 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  HatsTimeFrameModule   ·           -  ·          -  ·    1287099  ·        4.3 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  PullSplitFactory      ·           -  ·          -  ·    4535827  ·       15.1 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  PushSplitFactory      ·           -  ·          -  ·    4483113  ·       14.9 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  SplitsCreator         ·           -  ·          -  ·    1487532  ·          5 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  SplitsCreatorFactory  ·           -  ·          -  ·     526836  ·        1.8 %  ·          -  │
-·························|··············|·············|·············|···············|··············
-|  SplitsWarehouse       ·           -  ·          -  ·    3934655  ·       13.1 %  ·          -  │
 ·------------------------|--------------|-------------|-------------|---------------|-------------·
